+++ R/get_dep.R
@@ -1,27 +1,8 @@
-#' One type of dependencies of a package
-#'
-#' \code{get_dep} returns, in a vector, one type of dependencies of a package directly
-#' @param name String, name of the package
-#' @param type One of the following dependency words: "Depends", "Imports", "LinkingTo", "Suggests", "Reverse depends", "Reverse imports", "Reverse linking to", "Reverse suggests", up to letter case and space replaced by underscore
-#' @param scrape Boolean. If 'TRUE' (default), the page of the package will be scraped. If 'FALSE', tools::CRAN_package_db() will be used. Whether the argument equals 'TRUE' or 'FALSE' should not affect the output, but only the time taken. Usually, the former is faster than the latter for a single package.
-#' @return A string vector of dependencies
-#' @examples
-#' get_dep("dplyr", "Imports")
-#' get_dep("MASS", "Depends", TRUE) # FALSE will give same result
-#' @seealso \code{\link{get_dep_df}} for multiple types of dependencies, and \code{\link{get_dep_all_packages}} for the dependencies of all CRAN packages
+#' @rdname get_dep
 #' @export
-get_dep <- function(name, type, scrape = TRUE) {
-    type <- check_dep_word(type)
-    if (length(type) != 1L) {
-        stop("get_dep: type has to be one dependency word only.");
-    }
-    if (scrape) {
-        str0 <- get_dep_str(html_text_vec(cran_url(name)), type)
-    } else {
-        df0 <- as.data.frame(tools::CRAN_package_db(), stringsAsFactors = FALSE)
-        str0 <- df0[df0$Package == name, type][1L] # some packages have multiple rows
-    }
-    unique(get_dep_vec(str0))
+get_dep_df <- function(name, type, scrape = TRUE) {
+    .Deprecated("get_dep")
+    get_dep(name, type, scrape)
 }
 
 #' @rdname get_dep
@@ -33,32 +14,32 @@ get_dep_all <- function(name, type, scrape = TRUE) {
 
 #' Multiple types of dependencies
 #'
-#' \code{get_dep_df} returns a data frame of multiple types of dependencies of a package
+#' \code{get_dep} returns a data frame of multiple types of dependencies of a package
 #' @param name String, name of the package
-#' @param types A character vector that contains one or more of the following dependency words: "Depends", "Imports", "LinkingTo", "Suggests", "Reverse depends", "Reverse imports", "Reverse linking to", "Reverse suggests", up to letter case and space replaced by underscore. Alternatively, if 'types = "all"', all eight dependencies will be obtained.
+#' @param type A character vector that contains one or more of the following dependency words: "Depends", "Imports", "LinkingTo", "Suggests", "Enhances", "Reverse depends", "Reverse imports", "Reverse linking to", "Reverse suggests", "Reverse enhances", up to letter case and space replaced by underscore. Alternatively, if 'type = "all"', all ten dependencies will be obtained.
 #' @param scrape Boolean. If 'TRUE' (default), the page of the package will be scraped. If 'FALSE', tools::CRAN_package_db() will be used. Whether the argument equals 'TRUE' or 'FALSE' should not affect the output, but only the time taken. Usually, the former is faster than the latter for a single package.
 #' @importFrom tools CRAN_package_db
 #' @importFrom stringr str_detect
 #' @return A data frame of dependencies
 #' @examples
-#' get_dep_df("dplyr", c("Imports", "Depends"))
-#' get_dep_df("MASS", c("Suggests", "Depends", "Imports"), TRUE) # FALSE will give same result
-#' @seealso \code{\link{get_dep}} for getting one type of dependencies in a character vector, and \code{\link{get_dep_all_packages}} for the dependencies of all CRAN packages, and \code{\link{get_graph_all_packages}} for obtaining directly a network of dependencies as an igraph object
+#' get_dep("dplyr", c("Imports", "Depends"))
+#' get_dep("MASS", c("Suggests", "Depends", "Imports"), TRUE) # FALSE will give same result
+#' @seealso \code{\link{get_dep_all_packages}} for the dependencies of all CRAN packages, and \code{\link{get_graph_all_packages}} for obtaining directly a network of dependencies as an igraph object
 #' @export
-get_dep_df <- function(name, types, scrape = TRUE) {
-    types <- check_dep_word(types)
+get_dep <- function(name, type, scrape = TRUE) {
+    type <- check_dep_word(type)
     l0 <- list()
     if (scrape) {
         html0 <- html_text_vec(cran_url(name))
-        for (i in seq_along(types)) {
-            typei <- types[i]
+        for (i in seq_along(type)) {
+            typei <- type[i]
             v0 <- get_dep_vec(get_dep_str(html0, typei))
             l0[[i]] <- data.frame(from = name, to = v0, type = typei, stringsAsFactors = FALSE)
         }
     } else {
         df1 <- as.data.frame(tools::CRAN_package_db())
-        for (i in seq_along(types)) {
-            typei <- types[i]
+        for (i in seq_along(type)) {
+            typei <- type[i]
             str0 <- df1[df1$Package == name, typei]
             v0 <- get_dep_vec(str0[1L]) # some packages have multiple rows
             l0[[i]] <- data.frame(from = name, to = v0, type = typei, stringsAsFactors = FALSE)
@@ -92,7 +73,7 @@ reshape_dep <- function(x, names) {
 #'
 #' \code{get_dep_all_packages} returns the data frame of dependencies of all packages currently available on CRAN.
 #'
-#' Unlike \code{get_dep} and \code{get_dep_df}, there is no boolean argument `scrape`, as it is much faster to obtain the dependencies of all packages via `tools::CRAN_package_db()`.
+#' Unlike \code{get_dep}, there is no boolean argument `scrape`, as it is much faster to obtain the dependencies of all packages via `tools::CRAN_package_db()`.
 #' @importFrom tools CRAN_package_db
 #' @importFrom dplyr bind_rows
 #' @return A data frame of dependencies of all CRAN packages
@@ -100,7 +81,7 @@ reshape_dep <- function(x, names) {
 #' \dontrun{
 #' df.cran <- get_dep_all_packages()
 #' }
-#' @seealso \code{\link{get_dep}} for getting one type of dependencies in a character vector, \code{\link{get_dep_df}} for multiple types of dependencies, and \code{\link{get_graph_all_packages}} for obtaining directly a network of dependencies as an igraph object
+#' @seealso \code{\link{get_dep}} for multiple types of dependencies, and \code{\link{get_graph_all_packages}} for obtaining directly a network of dependencies as an igraph object
 #' @export
 get_dep_all_packages <- function() {
     df0 <- as.data.frame(tools::CRAN_package_db(), stringsAsFactors = FALSE)
@@ -111,6 +92,7 @@ get_dep_all_packages <- function() {
             depends = reshape_dep(df0$Depends, pkgnames),
             `linking to` = reshape_dep(df0$LinkingTo, pkgnames),
             suggests = reshape_dep(df0$Suggests, pkgnames),
+            enhances = reshape_dep(df0$Enhances, pkgnames),
             .id = "type"
         ),
         `TRUE` = dplyr::bind_rows(
@@ -118,6 +100,7 @@ get_dep_all_packages <- function() {
             depends = reshape_dep(df0$`Reverse depends`, pkgnames),
             `linking to` = reshape_dep(df0$`Reverse linking to`, pkgnames),
             suggests = reshape_dep(df0$`Reverse suggests`, pkgnames),
+            enhances = reshape_dep(df0$`Reverse enhances`, pkgnames),
             .id = "type"
         ),
         .id = "reverse"